+++ Sources/xcproj/PBXFileReference.swift
@@ -51,8 +51,8 @@ final public class PBXFileReference: PBXFileElement {
     }
 
     public override func isEqual(to object: PBXObject) -> Bool {
-        guard super.isEqual(to: self),
-            let rhs = object as? PBXFileReference else {
+        guard let rhs = object as? PBXFileReference,
+            super.isEqual(to: rhs) else {
                 return false
         }
         let lhs = self