+++ Sources/xcproj/PBXGroup.swift
@@ -41,8 +41,8 @@ final public class PBXGroup: PBXFileElement {
     }
 
     public override func isEqual(to object: PBXObject) -> Bool {
-        guard super.isEqual(to: self),
-            let rhs = object as? PBXGroup else {
+        guard let rhs = object as? PBXGroup,
+            super.isEqual(to: rhs) else {
                 return false
         }
         let lhs = self