+++ Sources/xcproj/XCVersionGroup.swift
@@ -31,8 +31,8 @@ final public class XCVersionGroup: PBXFileElement {
     }
 
     public override func isEqual(to object: PBXObject) -> Bool {
-        guard super.isEqual(to: self),
-            let rhs = object as? XCVersionGroup else {
+        guard let rhs = object as? XCVersionGroup,
+            super.isEqual(to: rhs) else {
                 return false
         }
         let lhs = self