+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Changed
 - Support for `XCConfig` project-relative includes https://github.com/xcodeswift/xcproj/pull/238 by @briantkelley
 
+### Fixed
+- `PBXObject.isEqual(to:)` overrides correctly call super https://github.com/xcodeswift/xcproj/pull/239 by @briantkelley
+
 ## 4.1.0
 
 ### Added